# docs/swagger.yaml
definitions:
  go-playground_internal_database_models.Author:
    properties:
      books:
        items:
          $ref: '#/definitions/go-playground_internal_database_models.Book'
        type: array
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      updatedAt:
        type: string
    required:
    - firstname
    - lastname
    type: object
  go-playground_internal_database_models.Book:
    properties:
      authorID:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      published_date:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    required:
    - published_date
    - title
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
  /authors:
    get:
      description: List all authors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go-playground_internal_database_models.Author'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List authors
      tags:
      - authors
    post:
      consumes:
      - application/json
      description: Create a new author
      parameters:
      - description: Author object
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/go-playground_internal_database_models.Author'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/go-playground_internal_database_models.Author'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create author
      tags:
      - authors
swagger: "2.0"